+++ spec/requests/project_information_controller_spec.rb
@@ -1,7 +1,7 @@
 require "rails_helper"
 
 RSpec.describe ProjectInformationController, type: :request do
-  let(:user) { User.create!(email: "user@education.gov.uk") }
+  let(:user) { create(:user, :team_leader) }
 
   before do
     mock_successful_authentication(user.email)
@@ -25,7 +25,43 @@ RSpec.describe ProjectInformationController, type: :request do
     end
 
     it "returns a successful response" do
-      expect(subject).to have_http_status :success
+      expect(perform_request).to have_http_status :success
+    end
+
+    describe "links to assign users to project roles" do
+      let(:change_links) do
+        [
+          project_assign_team_lead_path(project),
+          project_assign_regional_delivery_officer_path(project),
+          project_assign_caseworker_path(project)
+        ]
+      end
+
+      before { perform_request }
+
+      subject { response.body }
+
+      context "when team leader" do
+        it "has change links" do
+          expect(subject).to include(*change_links)
+        end
+      end
+
+      context "when regional delivery officer" do
+        let(:user) { create(:user, :regional_delivery_officer) }
+
+        it "does not have change links" do
+          expect(subject).to_not include(*change_links)
+        end
+      end
+
+      context "when caseworker" do
+        let(:user) { create(:user, :caseworker) }
+
+        it "does not have change links" do
+          expect(subject).to_not include(*change_links)
+        end
+      end
     end
   end
 end